+++ transporter/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* Added support for S3 as a storage backend (#[11](https://github.com/stjude-rust-labs/planetary/pull/11)).
 * Added signal handling ([#9](https://github.com/stjude-rust-labs/planetary/pull/9)).
 * Initial implementation of inputs and outputs using `cloud::copy` ([#4](https://github.com/stjude-rust-labs/planetary/pull/4)).